moves show function from main to format
